[FEAT] Booking
- set input upload to general component
- set input upload for passport & izin
- update upload file
