+++ src/components/molecules/Form/FormBooking/FormBooking.jsx
@@ -1,9 +1,8 @@
-import { InputDate, InputLabel, InputText } from '@/components/atoms';
+import { InputDate, InputLabel, InputText, InputUpload } from '@/components/atoms';
 import { setMaxDateOfBirth } from '@/utils/helpers';
 import { useForm, Controller } from 'react-hook-form';
 import { BsPlus } from 'react-icons/bs';
 import { InputSelectPersonTitle } from '../../InputSelect/InputSelectPersonTitle/InputSelectPersonTitle';
-import { InputUploadVisa } from '../../InputUpload/InputUploadVisa/InputUploadVisa';
 
 export const FormBooking = () => {
 	const { control, handleSubmit, setValue, setError } = useForm({
@@ -96,13 +95,21 @@ export const FormBooking = () => {
 									<InputDate {...field} label="Tanggal Lahir" maxDate={setMaxDateOfBirth(17)} error={error} />
 								)}
 							/>
+
+							<Controller
+								name="seat"
+								control={control}
+								render={({ field, fieldState: { error } }) => <InputText {...field} label="Pilih Nomor Kursi" error={error} />}
+							/>
 						</div>
 
 						<div className="space-y-3">
 							<InputLabel text="Upload Dokumen" />
 							<div className="p-4 border rounded-md space-y-4">
-								<div className="grid md:grid-cols-3">
-									<InputUploadVisa label="Upload Visa" helper="" onUploaded={(fileURL) => setValue('visa', fileURL)} />
+								<div className="grid md:grid-cols-3 gap-4">
+									<InputUpload name="visa" label="Upload Visa" onUploaded={(fileURL) => setValue('visa', fileURL)} />
+									<InputUpload name="passport" label="Upload Passport" onUploaded={(fileURL) => setValue('passport', fileURL)} />
+									<InputUpload name="izin" label="Upload Izin" onUploaded={(fileURL) => setValue('izin', fileURL)} />
 								</div>
 							</div>
 						</div>
