+++ src/store/transaction/transaction.reducer.js
@@ -1,4 +1,4 @@
-import { REQUEST_SAVE_TRANSACTION_TEMP_DATA, REQUEST_UPLOAD_DOCUMENT, RESPONSE_UPLOAD_DOCUMENT } from './transaction.types';
+import { SET_TRANSACTION_TEMP_DATA, REQUEST_UPLOAD_DOCUMENT, RESPONSE_UPLOAD_DOCUMENT } from './transaction.types';
 
 const initialState = {
 	transactionTempData: null,
@@ -8,7 +8,7 @@ const initialState = {
 
 export default function reducer(state = initialState, { type, payload }) {
 	switch (type) {
-		case REQUEST_SAVE_TRANSACTION_TEMP_DATA:
+		case SET_TRANSACTION_TEMP_DATA:
 			return {
 				...state,
 				transactionTempData: payload.data
